requirements.txt
pandas>=2.0
numpy>=1.24
pyyaml>=6.0
sqlalchemy>=2.0
psycopg2-binary>=2.9
scipy>=1.10
requests>=2.28
